+++ src/werkzeug/debug/tbtools.py
@@ -1,5 +1,5 @@
-import codecs
-import inspect
+import itertools
+import linecache
 import os
 import re
 import sys
@@ -7,21 +7,12 @@ import sysconfig
 import traceback
 import typing as t
 from html import escape
-from tokenize import TokenError
-from types import CodeType
-from types import TracebackType
 
-from .._internal import _to_str
 from ..utils import cached_property
 from .console import Console
 
-_coding_re = re.compile(br"coding[:=]\s*([-\w.]+)")
-_line_re = re.compile(br"^(.*?)$", re.MULTILINE)
-_funcdef_re = re.compile(r"^(\s*def\s)|(.*(?<!\w)lambda(:|\s))|^(\s*@)")
-
 HEADER = """\
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
-  "http://www.w3.org/TR/html4/loose.dtd">
+<!DOCTYPE html">
 <html>
   <head>
     <title>%(title)s // Werkzeug Debugger</title>
@@ -34,8 +25,7 @@ HEADER = """\
         href="?__debugger__=yes&amp;cmd=resource&amp;f=console.png">
     <script src="?__debugger__=yes&amp;cmd=resource&amp;f=debugger.js"></script>
     <script type="text/javascript">
-      var TRACEBACK = %(traceback_id)d,
-          CONSOLE_MODE = %(console)s,
+      var CONSOLE_MODE = %(console)s,
           EVALEX = %(evalex)s,
           EVALEX_TRUSTED = %(evalex_trusted)s,
           SECRET = "%(secret)s";
@@ -44,6 +34,7 @@ HEADER = """\
   <body style="background-color: #fff">
     <div class="debugger">
 """
+
 FOOTER = """\
       <div class="footer">
         Brought to you by <strong class="arthur">DON'T PANIC</strong>, your
@@ -132,336 +123,244 @@ FRAME_HTML = """\
 </div>
 """
 
-SOURCE_LINE_HTML = """\
-<tr class="%(classes)s">
-  <td class=lineno>%(lineno)s</td>
-  <td>%(code)s</td>
-</tr>
-"""
 
-
-def render_console_html(secret: str, evalex_trusted: bool = True) -> str:
-    return CONSOLE_HTML % {
-        "evalex": "true",
-        "evalex_trusted": "true" if evalex_trusted else "false",
-        "console": "true",
-        "title": "Console",
-        "secret": secret,
-        "traceback_id": -1,
-    }
-
-
-def get_current_traceback(
-    ignore_system_exceptions: bool = False,
-    show_hidden_frames: bool = False,
+def _process_traceback(
+    exc: BaseException,
+    te: t.Optional[traceback.TracebackException] = None,
+    *,
     skip: int = 0,
-) -> "Traceback":
-    """Get the current exception info as `Traceback` object.  Per default
-    calling this method will reraise system exceptions such as generator exit,
-    system exit or others.  This behavior can be disabled by passing `False`
-    to the function as first parameter.
-    """
-    info = t.cast(
-        t.Tuple[t.Type[BaseException], BaseException, TracebackType], sys.exc_info()
-    )
-    exc_type, exc_value, tb = info
+    hide: bool = True,
+) -> traceback.TracebackException:
+    if te is None:
+        te = traceback.TracebackException.from_exception(exc, lookup_lines=False)
 
-    if ignore_system_exceptions and exc_type in {
-        SystemExit,
-        KeyboardInterrupt,
-        GeneratorExit,
-    }:
-        raise
-    for _ in range(skip):
-        if tb.tb_next is None:
+    # Get the frames the same way StackSummary.extract did, in order
+    # to match each frame with the FrameSummary to augment.
+    frame_gen = traceback.walk_tb(exc.__traceback__)
+    limit = getattr(sys, "tracebacklimit", None)
+
+    if limit is not None:
+        if limit < 0:
+            limit = 0
+
+        frame_gen = itertools.islice(frame_gen, limit)
+
+    if skip:
+        frame_gen = itertools.islice(frame_gen, skip, None)
+        del te.stack[:skip]
+
+    new_stack: t.List[DebugFrameSummary] = []
+    hidden = False
+
+    # Match each frame with the FrameSummary that was generated.
+    # Hide frames using Paste's __traceback_hide__ rules. Replace
+    # all visible FrameSummary with DebugFrameSummary.
+    for (f, _), fs in zip(frame_gen, te.stack):
+        if hide:
+            hide_value = f.f_locals.get("__traceback_hide__", False)
+
+            if hide_value in {"before", "before_and_this"}:
+                new_stack = []
+                hidden = False
+
+                if hide_value == "before_and_this":
+                    continue
+            elif hide_value in {"reset", "reset_and_this"}:
+                hidden = False
+
+                if hide_value == "reset_and_this":
+                    continue
+            elif hide_value in {"after", "after_and_this"}:
+                hidden = True
+
+                if hide_value == "after_and_this":
+                    continue
+            elif hide_value or hidden:
+                continue
+
+        new_stack.append(
+            DebugFrameSummary(
+                filename=fs.filename,
+                lineno=fs.lineno,
+                name=fs.name,
+                locals=f.f_locals,
+                globals=f.f_globals,
+            )
+        )
+
+    # The codeop module is used to compile code from the interactive
+    # debugger. Hide any codeop frames from the bottom of the traceback.
+    while new_stack:
+        module = new_stack[0].global_ns.get("__name__")
+
+        if module is None:
+            module = new_stack[0].local_ns.get("__name__")
+
+        if module == "codeop":
+            del new_stack[0]
+        else:
             break
-        tb = tb.tb_next
-    tb = Traceback(exc_type, exc_value, tb)
-    if not show_hidden_frames:
-        tb.filter_hidden_frames()
-    return tb
+
+    te.stack[:] = new_stack
+
+    if te.__context__:
+        context_exc = t.cast(BaseException, exc.__context__)
+        te.__context__ = _process_traceback(context_exc, te.__context__, hide=hide)
+
+    if te.__cause__:
+        cause_exc = t.cast(BaseException, exc.__cause__)
+        te.__cause__ = _process_traceback(cause_exc, te.__cause__, hide=hide)
+
+    return te
 
 
-class Line:
-    """Helper for the source renderer."""
-
-    __slots__ = ("lineno", "code", "in_frame", "current")
-
-    def __init__(self, lineno: int, code: str) -> None:
-        self.lineno = lineno
-        self.code = code
-        self.in_frame = False
-        self.current = False
-
-    @property
-    def classes(self) -> t.List[str]:
-        rv = ["line"]
-        if self.in_frame:
-            rv.append("in-frame")
-        if self.current:
-            rv.append("current")
-        return rv
-
-    def render(self) -> str:
-        return SOURCE_LINE_HTML % {
-            "classes": " ".join(self.classes),
-            "lineno": self.lineno,
-            "code": escape(self.code),
-        }
-
-
-class Traceback:
-    """Wraps a traceback."""
+class DebugTraceback:
+    __slots__ = ("_te", "_cache_all_tracebacks", "_cache_all_frames")
 
     def __init__(
         self,
-        exc_type: t.Type[BaseException],
-        exc_value: BaseException,
-        tb: TracebackType,
+        exc: BaseException,
+        te: t.Optional[traceback.TracebackException] = None,
+        *,
+        skip: int = 0,
+        hide: bool = True,
     ) -> None:
-        self.exc_type = exc_type
-        self.exc_value = exc_value
-        self.tb = tb
+        self._te = _process_traceback(exc, te, skip=skip, hide=hide)
 
-        exception_type = exc_type.__name__
-        if exc_type.__module__ not in {"builtins", "__builtin__", "exceptions"}:
-            exception_type = f"{exc_type.__module__}.{exception_type}"
-        self.exception_type = exception_type
+    def __str__(self) -> str:
+        return f"<{type(self).__name__} {self._te}>"
 
-        self.groups = []
-        memo = set()
-        while True:
-            self.groups.append(Group(exc_type, exc_value, tb))
-            memo.add(id(exc_value))
-            exc_value = exc_value.__cause__ or exc_value.__context__  # type: ignore
-            if exc_value is None or id(exc_value) in memo:
-                break
-            exc_type = type(exc_value)
-            tb = exc_value.__traceback__  # type: ignore
-        self.groups.reverse()
-        self.frames = [frame for group in self.groups for frame in group.frames]
+    @cached_property
+    def all_tracebacks(
+        self,
+    ) -> t.List[t.Tuple[t.Optional[str], traceback.TracebackException]]:
+        out = []
+        current = self._te
 
-    def filter_hidden_frames(self) -> None:
-        """Remove the frames according to the paste spec."""
-        for group in self.groups:
-            group.filter_hidden_frames()
+        while current is not None:
+            if current.__cause__ is not None:
+                chained_msg = (
+                    "The above exception was the direct cause of the"
+                    " following exception"
+                )
+                chained_exc = current.__cause__
+            elif current.__context__ is not None and not current.__suppress_context__:
+                chained_msg = (
+                    "During handling of the above exception, another"
+                    " exception occurred"
+                )
+                chained_exc = current.__context__
+            else:
+                chained_msg = None
+                chained_exc = None
 
-        self.frames[:] = [frame for group in self.groups for frame in group.frames]
+            out.append((chained_msg, current))
+            current = chained_exc
 
-    @property
-    def is_syntax_error(self) -> bool:
-        """Is it a syntax error?"""
-        return isinstance(self.exc_value, SyntaxError)
+        return out
 
-    @property
-    def exception(self) -> str:
-        """String representation of the final exception."""
-        return self.groups[-1].exception
+    @cached_property
+    def all_frames(self) -> t.List["DebugFrameSummary"]:
+        return [
+            f for _, te in self.all_tracebacks for f in te.stack  # type: ignore[misc]
+        ]
 
-    def log(self, logfile: t.Optional[t.IO[str]] = None) -> None:
-        """Log the ASCII traceback into a file object."""
-        if logfile is None:
-            logfile = sys.stderr
-        tb = f"{self.plaintext.rstrip()}\n"
-        logfile.write(tb)
+    def render_traceback_text(self) -> str:
+        return "".join(self._te.format())
 
-    def render_summary(self, include_title: bool = True) -> str:
-        """Render the traceback for the interactive console."""
-        title = ""
-        classes = ["traceback"]
-        if not self.frames:
-            classes.append("noframe-traceback")
-            frames = []
+    def render_traceback_html(self, include_title: bool = True) -> str:
+        library_frames = [f.is_library for f in self.all_frames]
+        mark_library = 0 < sum(library_frames) < len(library_frames)
+        rows = []
+
+        if not library_frames:
+            classes = "traceback noframe-traceback"
         else:
-            library_frames = sum(frame.is_library for frame in self.frames)
-            mark_lib = 0 < library_frames < len(self.frames)
-            frames = [group.render(mark_lib=mark_lib) for group in self.groups]
+            classes = "traceback"
+
+            for msg, current in reversed(self.all_tracebacks):
+                row_parts = []
+
+                if msg is not None:
+                    row_parts.append(f'<li><div class="exc-divider">{msg}:</div>')
+
+                for frame in current.stack:
+                    frame = t.cast(DebugFrameSummary, frame)
+                    info = f' title="{escape(frame.info)}"' if frame.info else ""
+                    row_parts.append(f"<li{info}>{frame.render_html(mark_library)}")
+
+                rows.append("\n".join(row_parts))
+
+        is_syntax_error = issubclass(self._te.exc_type, SyntaxError)
 
         if include_title:
-            if self.is_syntax_error:
+            if is_syntax_error:
                 title = "Syntax Error"
             else:
                 title = "Traceback <em>(most recent call last)</em>:"
-
-        if self.is_syntax_error:
-            description = f"<pre class=syntaxerror>{escape(self.exception)}</pre>"
         else:
-            description = f"<blockquote>{escape(self.exception)}</blockquote>"
+            title = ""
+
+        exc_full = escape("".join(self._te.format_exception_only()))
+
+        if is_syntax_error:
+            description = f"<pre class=syntaxerror>{exc_full}</pre>"
+        else:
+            description = f"<blockquote>{exc_full}</blockquote>"
 
         return SUMMARY_HTML % {
-            "classes": " ".join(classes),
-            "title": f"<h3>{title if title else ''}</h3>",
-            "frames": "\n".join(frames),
+            "classes": classes,
+            "title": f"<h3>{title}</h3>",
+            "frames": "\n".join(rows),
             "description": description,
         }
 
-    def render_full(
-        self,
-        evalex: bool = False,
-        secret: t.Optional[str] = None,
-        evalex_trusted: bool = True,
+    def render_debugger_html(
+        self, evalex: bool, secret: str, evalex_trusted: bool
     ) -> str:
-        """Render the Full HTML page with the traceback info."""
-        exc = escape(self.exception)
+        exc_lines = list(self._te.format_exception_only())
+        plaintext = "".join(self._te.format())
         return PAGE_HTML % {
             "evalex": "true" if evalex else "false",
             "evalex_trusted": "true" if evalex_trusted else "false",
             "console": "false",
-            "title": exc,
-            "exception": exc,
-            "exception_type": escape(self.exception_type),
-            "summary": self.render_summary(include_title=False),
-            "plaintext": escape(self.plaintext),
-            "plaintext_cs": re.sub("-{2,}", "-", self.plaintext),
-            "traceback_id": self.id,
+            "title": exc_lines[0],
+            "exception": escape("".join(exc_lines)),
+            "exception_type": escape(self._te.exc_type.__name__),
+            "summary": self.render_traceback_html(include_title=False),
+            "plaintext": escape(plaintext),
+            "plaintext_cs": re.sub("-{2,}", "-", plaintext),
             "secret": secret,
         }
 
-    @cached_property
-    def plaintext(self) -> str:
-        return "\n".join([group.render_text() for group in self.groups])
 
-    @property
-    def id(self) -> int:
-        return id(self)
-
-
-class Group:
-    """A group of frames for an exception in a traceback. If the
-    exception has a ``__cause__`` or ``__context__``, there are multiple
-    exception groups.
+class DebugFrameSummary(traceback.FrameSummary):
+    """A :class:`traceback.FrameSummary` that can evaluate code in the
+    frame's namespace.
     """
 
-    def __init__(
-        self,
-        exc_type: t.Type[BaseException],
-        exc_value: BaseException,
-        tb: TracebackType,
-    ) -> None:
-        self.exc_type = exc_type
-        self.exc_value = exc_value
-        self.info = None
-        if exc_value.__cause__ is not None:
-            self.info = (
-                "The above exception was the direct cause of the following exception"
-            )
-        elif exc_value.__context__ is not None:
-            self.info = (
-                "During handling of the above exception, another exception occurred"
-            )
-
-        self.frames = []
-        while tb is not None:
-            self.frames.append(Frame(exc_type, exc_value, tb))
-            tb = tb.tb_next  # type: ignore
-
-    def filter_hidden_frames(self) -> None:
-        # An exception may not have a traceback to filter frames, such
-        # as one re-raised from ProcessPoolExecutor.
-        if not self.frames:
-            return
-
-        new_frames: t.List[Frame] = []
-        hidden = False
-
-        for frame in self.frames:
-            hide = frame.hide
-            if hide in ("before", "before_and_this"):
-                new_frames = []
-                hidden = False
-                if hide == "before_and_this":
-                    continue
-            elif hide in ("reset", "reset_and_this"):
-                hidden = False
-                if hide == "reset_and_this":
-                    continue
-            elif hide in ("after", "after_and_this"):
-                hidden = True
-                if hide == "after_and_this":
-                    continue
-            elif hide or hidden:
-                continue
-            new_frames.append(frame)
-
-        # if we only have one frame and that frame is from the codeop
-        # module, remove it.
-        if len(new_frames) == 1 and self.frames[0].module == "codeop":
-            del self.frames[:]
-
-        # if the last frame is missing something went terrible wrong :(
-        elif self.frames[-1] in new_frames:
-            self.frames[:] = new_frames
-
-    @property
-    def exception(self) -> str:
-        """String representation of the exception."""
-        buf = traceback.format_exception_only(self.exc_type, self.exc_value)
-        rv = "".join(buf).strip()
-        return _to_str(rv, "utf-8", "replace")
-
-    def render(self, mark_lib: bool = True) -> str:
-        out = []
-        if self.info is not None:
-            out.append(f'<li><div class="exc-divider">{self.info}:</div>')
-        for frame in self.frames:
-            title = f' title="{escape(frame.info)}"' if frame.info else ""
-            out.append(f"<li{title}>{frame.render(mark_lib=mark_lib)}")
-        return "\n".join(out)
-
-    def render_text(self) -> str:
-        out = []
-        if self.info is not None:
-            out.append(f"\n{self.info}:\n")
-        out.append("Traceback (most recent call last):")
-        for frame in self.frames:
-            out.append(frame.render_text())
-        out.append(self.exception)
-        return "\n".join(out)
-
-
-class Frame:
-    """A single frame in a traceback."""
+    __slots__ = (
+        "local_ns",
+        "global_ns",
+        "_cache_info",
+        "_cache_is_library",
+        "_cache_console",
+    )
 
     def __init__(
         self,
-        exc_type: t.Type[BaseException],
-        exc_value: BaseException,
-        tb: TracebackType,
+        *,
+        locals: t.Dict[str, t.Any],
+        globals: t.Dict[str, t.Any],
+        **kwargs: t.Any,
     ) -> None:
-        self.lineno = tb.tb_lineno
-        self.function_name = tb.tb_frame.f_code.co_name
-        self.locals = tb.tb_frame.f_locals
-        self.globals = tb.tb_frame.f_globals
+        super().__init__(locals=None, **kwargs)
+        self.local_ns = locals
+        self.global_ns = globals
 
-        fn = inspect.getsourcefile(tb) or inspect.getfile(tb)
-        if fn[-4:] in (".pyo", ".pyc"):
-            fn = fn[:-1]
-        # if it's a file on the file system resolve the real filename.
-        if os.path.isfile(fn):
-            fn = os.path.realpath(fn)
-
-        self.filename = os.fsdecode(fn)
-        self.module = self.globals.get("__name__", self.locals.get("__name__"))
-        self.loader = self.globals.get("__loader__", self.locals.get("__loader__"))
-        self.code = tb.tb_frame.f_code
-
-        # support for paste's traceback extensions
-        self.hide = self.locals.get("__traceback_hide__", False)
-        info = self.locals.get("__traceback_info__")
-        if info is not None:
-            info = _to_str(info, "utf-8", "replace")
-        self.info = info
-
-    def render(self, mark_lib: bool = True) -> str:
-        """Render a single frame in a traceback."""
-        return FRAME_HTML % {
-            "id": self.id,
-            "filename": escape(self.filename),
-            "lineno": self.lineno,
-            "function_name": escape(self.function_name),
-            "lines": self.render_line_context(),
-            "library": "library" if mark_lib and self.is_library else "",
-        }
+    @cached_property
+    def info(self) -> t.Optional[str]:
+        return self.local_ns.get("__traceback_info__")
 
     @cached_property
     def is_library(self) -> bool:
@@ -470,131 +369,54 @@ class Frame:
             for path in sysconfig.get_paths().values()
         )
 
-    def render_text(self) -> str:
-        return (
-            f'  File "{self.filename}", line {self.lineno}, in {self.function_name}\n'
-            f"    {self.current_line.strip()}"
-        )
+    @cached_property
+    def console(self) -> Console:
+        return Console(self.global_ns, self.local_ns)
 
-    def render_line_context(self) -> str:
-        before, current, after = self.get_context_lines()
-        rv = []
+    def eval(self, code: str) -> t.Any:
+        return self.console.eval(code)
+
+    def render_html(self, mark_library: bool) -> str:
+        context = 5
+        lines = linecache.getlines(self.filename)
+        line_idx = self.lineno - 1
+        start_idx = max(0, line_idx - context)
+        stop_idx = min(len(lines), line_idx + context + 1)
+        rendered_lines = []
 
         def render_line(line: str, cls: str) -> None:
             line = line.expandtabs().rstrip()
             stripped_line = line.strip()
             prefix = len(line) - len(stripped_line)
-            rv.append(
+            rendered_lines.append(
                 f'<pre class="line {cls}"><span class="ws">{" " * prefix}</span>'
                 f"{escape(stripped_line) if stripped_line else ' '}</pre>"
             )
 
-        for line in before:
-            render_line(line, "before")
-        render_line(current, "current")
-        for line in after:
-            render_line(line, "after")
+        if lines:
+            for line in lines[start_idx:line_idx]:
+                render_line(line, "before")
 
-        return "\n".join(rv)
+            render_line(lines[line_idx], "current")
 
-    def get_annotated_lines(self) -> t.List[Line]:
-        """Helper function that returns lines with extra information."""
-        lines = [Line(idx + 1, x) for idx, x in enumerate(self.sourcelines)]
+            for line in lines[line_idx + 1 : stop_idx]:
+                render_line(line, "after")
 
-        # find function definition and mark lines
-        if hasattr(self.code, "co_firstlineno"):
-            lineno = self.code.co_firstlineno - 1
-            while lineno > 0:
-                if _funcdef_re.match(lines[lineno].code):
-                    break
-                lineno -= 1
-            try:
-                offset = len(inspect.getblock([f"{x.code}\n" for x in lines[lineno:]]))
-            except TokenError:
-                offset = 0
-            for line in lines[lineno : lineno + offset]:
-                line.in_frame = True
+        return FRAME_HTML % {
+            "id": id(self),
+            "filename": escape(self.filename),
+            "lineno": self.lineno,
+            "function_name": escape(self.name),
+            "lines": "\n".join(rendered_lines),
+            "library": "library" if mark_library and self.is_library else "",
+        }
 
-        # mark current line
-        try:
-            lines[self.lineno - 1].current = True
-        except IndexError:
-            pass
 
-        return lines
-
-    def eval(self, code: t.Union[str, CodeType], mode: str = "single") -> t.Any:
-        """Evaluate code in the context of the frame."""
-        if isinstance(code, str):
-            code = compile(code, "<interactive>", mode)
-        return eval(code, self.globals, self.locals)
-
-    @cached_property
-    def sourcelines(self) -> t.List[str]:
-        """The sourcecode of the file as list of strings."""
-        # get sourcecode from loader or file
-        source = None
-        if self.loader is not None:
-            try:
-                if hasattr(self.loader, "get_source"):
-                    source = self.loader.get_source(self.module)
-                elif hasattr(self.loader, "get_source_by_code"):
-                    source = self.loader.get_source_by_code(self.code)
-            except Exception:
-                # we munch the exception so that we don't cause troubles
-                # if the loader is broken.
-                pass
-
-        if source is None:
-            try:
-                with open(self.filename, mode="rb") as f:
-                    source = f.read()
-            except OSError:
-                return []
-
-        # already str?  return right away
-        if isinstance(source, str):
-            return source.splitlines()
-
-        charset = "utf-8"
-        if source.startswith(codecs.BOM_UTF8):
-            source = source[3:]
-        else:
-            for idx, match in enumerate(_line_re.finditer(source)):
-                coding_match = _coding_re.search(match.group())
-                if coding_match is not None:
-                    charset = coding_match.group(1).decode("utf-8")
-                    break
-                if idx > 1:
-                    break
-
-        # on broken cookies we fall back to utf-8 too
-        charset = _to_str(charset)
-        try:
-            codecs.lookup(charset)
-        except LookupError:
-            charset = "utf-8"
-
-        return source.decode(charset, "replace").splitlines()
-
-    def get_context_lines(
-        self, context: int = 5
-    ) -> t.Tuple[t.List[str], str, t.List[str]]:
-        before = self.sourcelines[self.lineno - context - 1 : self.lineno - 1]
-        past = self.sourcelines[self.lineno : self.lineno + context]
-        return (before, self.current_line, past)
-
-    @property
-    def current_line(self) -> str:
-        try:
-            return self.sourcelines[self.lineno - 1]
-        except IndexError:
-            return ""
-
-    @cached_property
-    def console(self) -> Console:
-        return Console(self.globals, self.locals)
-
-    @property
-    def id(self) -> int:
-        return id(self)
+def render_console_html(secret: str, evalex_trusted: bool) -> str:
+    return CONSOLE_HTML % {
+        "evalex": "true",
+        "evalex_trusted": "true" if evalex_trusted else "false",
+        "console": "true",
+        "title": "Console",
+        "secret": secret,
+    }